gh8 simplified logging setup
